use int32 in pgx

+++ pgx/_src/gardner_chess_utils.py
@@ -2,8 +2,8 @@
 import jax
 import jax.numpy as jnp
 
-TO_MAP = -jnp.ones((25, 49), dtype=jnp.int8)
-PLANE_MAP = -jnp.ones((25, 25), dtype=jnp.int8)  # ignores underpromotions
+TO_MAP = -jnp.ones((25, 49), dtype=jnp.int32)
+PLANE_MAP = -jnp.ones((25, 25), dtype=jnp.int32)  # ignores underpromotions
 # underpromotions
 for from_ in range(25):
     if from_ % 5 != 3:  # 4th row in current player view
@@ -16,7 +16,7 @@ for from_ in range(25):
         # board index (flipped black view)
         # 5  0  5 10 15 20
         # 4  1  6 11 16 21
-        to = from_ + jnp.int8([+1, +6, -4])[dir_]
+        to = from_ + jnp.int32([+1, +6, -4])[dir_]
         if not (0 <= to < 25):
             continue
         TO_MAP = TO_MAP.at[from_, plane].set(to)
@@ -32,12 +32,12 @@ for from_ in range(25):
         c = c + dc[plane - 9]
         if r < 0 or r >= 5 or c < 0 or c >= 5:
             continue
-        to = jnp.int8(c * 5 + r)
+        to = jnp.int32(c * 5 + r)
         TO_MAP = TO_MAP.at[from_, plane].set(to)
-        PLANE_MAP = PLANE_MAP.at[from_, to].set(jnp.int8(plane))
+        PLANE_MAP = PLANE_MAP.at[from_, to].set(jnp.int32(plane))
 
 
-CAN_MOVE = -jnp.ones((7, 25, 16), jnp.int8)
+CAN_MOVE = -jnp.ones((7, 25, 16), jnp.int32)
 # usage: CAN_MOVE[piece, from_x, from_y]
 # CAN_MOVE[0, :, :] are all -1
 # Note that the board is not symmetric about the center (different from shogi)
@@ -55,7 +55,7 @@ for from_ in range(25):
         if r1 - r0 == 1 and jnp.abs(c1 - c0) <= 1:
             legal_dst.append(to)
     assert len(legal_dst) <= 6, f"{from_=}, {to=}, {legal_dst=}"
-    CAN_MOVE = CAN_MOVE.at[1, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[1, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 # KNIGHT
 for from_ in range(25):
     r0, c0 = from_ % 5, from_ // 5
@@ -67,7 +67,7 @@ for from_ in range(25):
         if jnp.abs(r1 - r0) == 2 and jnp.abs(c1 - c0) == 1:
             legal_dst.append(to)
     assert len(legal_dst) <= 8
-    CAN_MOVE = CAN_MOVE.at[2, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[2, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 # BISHOP
 for from_ in range(25):
     r0, c0 = from_ % 5, from_ // 5
@@ -79,7 +79,7 @@ for from_ in range(25):
         if jnp.abs(r1 - r0) == jnp.abs(c1 - c0):
             legal_dst.append(to)
     assert len(legal_dst) <= 8
-    CAN_MOVE = CAN_MOVE.at[3, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[3, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 # ROOK
 for from_ in range(25):
     r0, c0 = from_ % 5, from_ // 5
@@ -91,7 +91,7 @@ for from_ in range(25):
         if jnp.abs(r1 - r0) == 0 or jnp.abs(c1 - c0) == 0:
             legal_dst.append(to)
     assert len(legal_dst) <= 8
-    CAN_MOVE = CAN_MOVE.at[4, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[4, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 # QUEEN
 for from_ in range(25):
     r0, c0 = from_ % 5, from_ // 5
@@ -105,7 +105,7 @@ for from_ in range(25):
         if jnp.abs(r1 - r0) == jnp.abs(c1 - c0):
             legal_dst.append(to)
     assert len(legal_dst) <= 16
-    CAN_MOVE = CAN_MOVE.at[5, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[5, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 # KING
 for from_ in range(25):
     r0, c0 = from_ % 5, from_ // 5
@@ -117,11 +117,11 @@ for from_ in range(25):
         if (jnp.abs(r1 - r0) <= 1) and (jnp.abs(c1 - c0) <= 1):
             legal_dst.append(to)
     assert len(legal_dst) <= 8
-    CAN_MOVE = CAN_MOVE.at[6, from_, : len(legal_dst)].set(jnp.int8(legal_dst))
+    CAN_MOVE = CAN_MOVE.at[6, from_, : len(legal_dst)].set(jnp.int32(legal_dst))
 
 assert (CAN_MOVE[0, :, :] == -1).all()
 
-CAN_MOVE_ANY = -jnp.ones((25, 24), jnp.int8)
+CAN_MOVE_ANY = -jnp.ones((25, 24), jnp.int32)
 for from_ in range(25):
     legal_dst = []
     for i in range(16):
@@ -134,10 +134,10 @@ for from_ in range(25):
             legal_dst.append(to)
     assert len(legal_dst) <= 24
     CAN_MOVE_ANY = CAN_MOVE_ANY.at[from_, : len(legal_dst)].set(
-        jnp.int8(legal_dst)
+        jnp.int32(legal_dst)
     )
 
-BETWEEN = -jnp.ones((25, 25, 3), dtype=jnp.int8)
+BETWEEN = -jnp.ones((25, 25, 3), dtype=jnp.int32)
 for from_ in range(25):
     for to in range(25):
         r0, c0 = from_ % 5, from_ // 5
@@ -159,7 +159,7 @@ for from_ in range(25):
                 break
             bet.append(c * 5 + r)
         assert len(bet) <= 3
-        BETWEEN = BETWEEN.at[from_, to, : len(bet)].set(jnp.int8(bet))
+        BETWEEN = BETWEEN.at[from_, to, : len(bet)].set(jnp.int32(bet))
 
 
 INIT_LEGAL_ACTION_MASK = jnp.zeros(25 * 49, dtype=jnp.bool_)
